backend/requirements.txt
flask
flask-cors
python-dotenv
google-generativeai
flask
flask-cors
requests